blob: fix type in spdk_blob_resize() declaration
The 'sz' parameter to spdk_blob_resize() did not match between the
declaration and definition; it should be uint64_t, since it contains a
number of clusters.

+++ include/spdk/blob.h
@@ -422,7 +422,8 @@ void spdk_bs_open_blob(struct spdk_blob_store *bs, spdk_blob_id blobid,
  * \param cb_arg Argument passed to function cb_fn.
  *
  */
-void spdk_blob_resize(struct spdk_blob *blob, size_t sz, spdk_blob_op_complete cb_fn, void *cb_arg);
+void spdk_blob_resize(struct spdk_blob *blob, uint64_t sz, spdk_blob_op_complete cb_fn,
+		      void *cb_arg);
 
 /**
  * Set blob as read only.
